Generate a different state value for each auth request
A 'state' value is generated and saved (in local storage) before
redirecting the user to the auth server. After user has granted access,
this value is retrieved from the local storage and validated against the
authorization response. This prevents CSRF attacks, and will help
associating the response with the initial request if needed to identify
the platform that issued the the access token.